+++ lua/cmp_obsidian_new.lua
@@ -2,6 +2,7 @@ local abc = require "obsidian.abc"
 local completion = require "obsidian.completion.refs"
 local obsidian = require "obsidian"
 local log = require "obsidian.log"
+local LinkStyle = require("obsidian.config").LinkStyle
 
 ---@class cmp_obsidian_new.Source : obsidian.ABC
 local source = abc.new_class()
@@ -35,7 +36,7 @@ source.complete = function(_, request, callback)
   end
 
   if can_complete and search ~= nil and #search >= client.opts.completion.min_chars then
-    local new_title, new_id, path, rel_path
+    local new_title, new_id, path
     new_id = client:new_note_id(search)
     new_title, new_id, path = client:parse_title_id_path(search, new_id, dir)
 
@@ -43,39 +44,20 @@ source.complete = function(_, request, callback)
       return
     end
 
-    rel_path = client:vault_relative_path(path)
-    if rel_path == nil then
-      log.err("Failed to resolve path '%s' relative to vault root '%s'", path, client:vault_root())
-      return
-    end
-
-    if vim.endswith(rel_path, ".md") then
-      rel_path = string.sub(rel_path, 1, -4)
-    end
-
-    ---@type string, string
-    local sort_text, label, new_text
+    ---@type obsidian.config.LinkStyle, string, string
+    local link_style, sort_text
     if ref_type == completion.RefType.Wiki then
-      if client.opts.completion.use_path_only then
-        new_title = rel_path
-      elseif client.opts.completion.prepend_note_path then
-        new_title = rel_path .. "|" .. new_title
-      elseif client.opts.completion.prepend_note_id then
-        new_title = new_id .. "|" .. new_title
-      else
-        log.err "Invalid completion options"
-        return
-      end
+      link_style = LinkStyle.wiki
       sort_text = "[[" .. search
-      label = "Create: [[" .. new_title .. "]]"
-      new_text = "[[" .. new_title .. "]]"
     elseif ref_type == completion.RefType.Markdown then
+      link_style = LinkStyle.markdown
       sort_text = "[" .. search
-      label = "Create: [" .. new_title .. "](" .. rel_path .. ".md)"
-      new_text = "[" .. new_title .. "](" .. rel_path .. ".md)"
     else
       error "not implemented"
     end
+
+    local new_text = client:format_link(tostring(path), { label = new_title, link_style = link_style, id = new_id })
+    local label = "Create: " .. new_text
 
     local items = {
       {