+++ lua/obsidian/config.lua
@@ -162,6 +162,12 @@ config.LocationListOpts.default = function()
   }
 end
 
+---@enum obsidian.config.LinkStyle
+config.LinkStyle = {
+  wiki = "wiki",
+  markdown = "markdown",
+}
+
 ---@class obsidian.config.CompletionOpts
 ---@field nvim_cmp boolean
 ---@field min_chars integer